Refactor file IO code
