+++ MHWISaveEditor.vcxproj
@@ -22,6 +22,7 @@
     <ClInclude Include="data\EquipmentDB.h" />
     <ClInclude Include="gui\common\Notification.h" />
     <ClInclude Include="gui\common\WheelGuard.h" />
+    <ClInclude Include="utility\types.h" />
     <QtMoc Include="gui\equipment\equipmenteditortab.h" />
     <QtMoc Include="gui\equipment\equipmentslotview.h" />
     <ClInclude Include="data\SmithyDB.h" />
