+++ utility/read_bin_file.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <QFile>
+#include <QSaveFile>
 
 #include "../types/types.h"
 #include "../types/language.h"
@@ -18,50 +19,12 @@
 #include "../types/file/wp_dat_g.h"
 
 #include "system/paths.h"
-
-static u8* QByteArrayToU8(QByteArray arr, u8* dst, u32 size) {
-  u8* newdst = dst;
-  if (arr.length() != size) {
-    qWarning("Error: array size does not match desired size.");
-    return nullptr;
-  }
-
-  if (!newdst) newdst = (u8*)malloc(size);
-  if (!newdst) {
-    qWarning("Error allocating memory.");
-    return nullptr;
-  };
-  memcpy(newdst, arr.constData(), arr.length());
-
-  if (dst != newdst && dst) free(dst);
-  return newdst;
-}
-
-static u8* ReadEntireFileSize(const QString& path, u8* dst = nullptr, u32 *size = nullptr) {
-  QFile file(path);
-  if (!file.open(QIODevice::ReadOnly)) {
-    qWarning("Cannot open file: %s", qUtf8Printable(path));
-    return nullptr;
-  }
-
-  QByteArray blob = file.readAll();
-  u32 array_size = 0;
-  if (size) array_size = *size;
-  if (array_size == 0) array_size = blob.size();
-  if (size) *size = array_size;
-  file.close();
-
-  return QByteArrayToU8(blob, dst, array_size);
-}
-
-static u8* ReadEntireFile(const QString& path, u8* dst = nullptr, u32 size = 0) {
-  return ReadEntireFileSize(path, dst, &size);
-}
+#include "system/file_utils.h"
 
 static bool ReadMetaFile(gmd_meta* meta, const QString& path)
 {
   FreeMetaFile(meta);
-  gmd_header* header = (gmd_header*)ReadEntireFile(path);
+  gmd_header* header = (gmd_header*)FileUtils::ReadEntireFile(path);
   if (!header) {
     return false;
   }
@@ -71,7 +34,7 @@ static bool ReadMetaFile(gmd_meta* meta, const QString& path)
 
 static bool ReadMetaFile(itm_meta* meta, const QString& path) {
   FreeMetaFile(meta);
-  itm_header* header = (itm_header*)ReadEntireFile(path);
+  itm_header* header = (itm_header*)FileUtils::ReadEntireFile(path);
   if (!header) {
     return false;
   }
@@ -81,7 +44,7 @@ static bool ReadMetaFile(itm_meta* meta, const QString& path) {
 
 static bool ReadMetaFile(am_dat_meta* meta, const QString& path) {
   FreeMetaFile(meta);
-  am_dat_header* header = (am_dat_header*)ReadEntireFile(path);
+  am_dat_header* header = (am_dat_header*)FileUtils::ReadEntireFile(path);
   if (!header) {
     return false;
   }
@@ -91,7 +54,7 @@ static bool ReadMetaFile(am_dat_meta* meta, const QString& path) {
 
 static bool ReadMetaFile(eq_crt_meta* meta, const QString& path) {
   FreeMetaFile(meta);
-  eq_crt_header* header = (eq_crt_header*)ReadEntireFile(path);
+  eq_crt_header* header = (eq_crt_header*)FileUtils::ReadEntireFile(path);
   if (!header) {
     return false;
   }
@@ -101,7 +64,7 @@ static bool ReadMetaFile(eq_crt_meta* meta, const QString& path) {
 
 static bool ReadMetaFile(eq_cus_meta* meta, const QString& path) {
   FreeMetaFile(meta);
-  eq_cus_header* header = (eq_cus_header*)ReadEntireFile(path);
+  eq_cus_header* header = (eq_cus_header*)FileUtils::ReadEntireFile(path);
   if (!header) {
     return false;
   }
@@ -112,7 +75,7 @@ static bool ReadMetaFile(eq_cus_meta* meta, const QString& path) {
 static bool ReadMetaFile(rod_inse_meta* meta, const QString& path) {
   FreeMetaFile(meta);
   u32 size = 0;
-  rod_inse_header* header = (rod_inse_header*)ReadEntireFileSize(path, nullptr, &size);
+  rod_inse_header* header = (rod_inse_header*)FileUtils::ReadEntireFileSize(path, nullptr, &size);
   if (!header) {
     return false;
   }
@@ -123,7 +86,7 @@ static bool ReadMetaFile(rod_inse_meta* meta, const QString& path) {
 
 static bool ReadMetaFile(wp_dat_meta* meta, const QString& path) {
   FreeMetaFile(meta);
-  wp_dat_header* header = (wp_dat_header*)ReadEntireFile(path);
+  wp_dat_header* header = (wp_dat_header*)FileUtils::ReadEntireFile(path);
   if (!header) {
     return false;
   }
@@ -133,7 +96,7 @@ static bool ReadMetaFile(wp_dat_meta* meta, const QString& path) {
 
 static bool ReadMetaFile(wp_dat_g_meta* meta, const QString& path) {
   FreeMetaFile(meta);
-  wp_dat_g_header* header = (wp_dat_g_header*)ReadEntireFile(path);
+  wp_dat_g_header* header = (wp_dat_g_header*)FileUtils::ReadEntireFile(path);
   if (!header) {
     return false;
   }
@@ -143,7 +106,7 @@ static bool ReadMetaFile(wp_dat_g_meta* meta, const QString& path) {
 
 static bool ReadMetaFile(uct_meta* meta, const QString& path) {
   FreeMetaFile(meta);
-  uct_header* header = (uct_header*)ReadEntireFile(path);
+  uct_header* header = (uct_header*)FileUtils::ReadEntireFile(path);
   if (!header) {
     return false;
   }
